Create update scout screen
This screen first displays every scout in the database in two columns
and each name acts as a hyperlink to the second stage of the screen.
This second stage mirrors the add scout screen however the information
for that scout which is already in the database is drawn in the form.
When the changes are made, they are committed in database and the table
for that scout is renamed in the database.