Create delete scout screen
This screen first displays every scout in the database in two columns
and each name acts as a hyperlink to the second stage of the screen. The
second stage is to prompt the user to confirm whether to delete the
scout. If this is confirmed then the scout is deleted and the table for
that scout is dropped from the database.