Create update badge screen
This screen first displays every badge in the database in two columns
and each name acts as a hyperlink to the second stage of the screen.
This second stage mirrors the add badge screen however the information
for that badge which is already in the database is drawn in the form.
When the changes are made, they are committed in database.