Create delete badge screen
This screen first displays every badge in the database in two columns
and each name acts as a hyperlink to the second stage of the screen. The
second stage is to prompt the user to confirm whether to delete the
badge. If this is confirmed then the badge is deleted.