Create Email Addresses Report
This screen allows the user to select email addresses within the system
to be exported as comma separated values. The email addresses are
selected using a checkbox and then displayed in a pop-up window. This
allows the email addresses to be used in an email client. The screen is
accessible from the Reports section.