Create Badges Screen
This allows the user to enter the date when a scout achieved a certain
requirement of a badge. The dates are then saved into a database to be
recalled in the future. If a scout completes enough requirements to be
given a badge then the user is informed of this and the current date is
saved in the system as the date that the badge was awarded. This area is
accessible from the reports screen.